refactor(Computer): Computer 클래스 리팩토링
